+++ R/build_site.R
@@ -77,7 +77,7 @@ build_site <- function(path = ".", quiet = !interactive(), preview = TRUE, overr
   # Bypass certain downlit functions that produce unintented effects such
   # as linking function documentation.
   shimstem_file <- system.file("pkgdown", "shim.R", package = "sandpaper")
-  expected <- "62a291ba3386aad91e5ade20480ad7cf" 
+  expected <- "230853fec984d1a0e5766d3da79f1cea" 
   actual   <- tools::md5sum(shimstem_file)
   if (expected == actual) {
     # evaluate the shim in our namespace